Add index and update cart route

+++ src/Controller/CartController.php
@@ -8,7 +8,7 @@ use App\DTO\CartDto;
 use App\Entity\Cart;
 use App\Factory\CartFactory;
 use App\Repository\CartRepository;
-use App\Repository\ProductRepository;
+use App\Service\CartService;
 use Doctrine\ORM\EntityManagerInterface;
 use Nelmio\ApiDocBundle\Attribute\Security;
 use OpenApi\Attributes as OA;
@@ -19,6 +19,7 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Attribute\MapRequestPayload;
 use Symfony\Component\Routing\Attribute\Route;
 use Symfony\Component\Security\Http\Attribute\IsGranted;
+use Symfony\Component\Serializer\SerializerInterface;
 
 #[OA\Tag(name: 'Cart')]
 class CartController extends AbstractController
@@ -26,9 +27,22 @@ class CartController extends AbstractController
     public function __construct(
         private readonly CartRepository $cartRepository,
         private readonly CartFactory $cartFactory,
-        private readonly ProductRepository $productRepository,
         private readonly  EntityManagerInterface $entityManager,
+        private readonly SerializerInterface $serializer,
+        private readonly CartService $cartService,
     ) {
+    }
+
+    #[Route('/api/carts/{id}', name: 'cart_index', methods: ['POST'])]
+    public function index(#[MapEntity(id: 'id')] Cart $cart): JsonResponse
+    {
+        $errors = $this->cartService->checkPermission($cart->getUser(), $this->getUser(), $this->isGranted('ROLE_ADMIN'));
+        if ($errors !== null) {
+            return $errors;
+        }
+        $cartData = $this->serializer->serialize($cart, 'json', ['groups' => 'cart']);
+
+        return new JsonResponse($cartData, Response::HTTP_OK);
     }
 
     #[Route('/api/carts', name: 'cart_create', methods: ['POST'])]
@@ -41,20 +55,16 @@ class CartController extends AbstractController
             validationFailedStatusCode: Response::HTTP_BAD_REQUEST,
         )] CartDTO $cartDTO,
     ): JsonResponse {
-        foreach ($cartDTO->cartItems as $item) {
-            $product = $this->productRepository->find($item->productId);
-            if ($product === null) {
-                return new JsonResponse(['error' => 'Product not found. ID:' . $item->productId], Response::HTTP_BAD_REQUEST);
-            }
-            $item->product = $product;
-        }
         $user = $this->getUser();
         if ($this->cartRepository->findByUser($user) !== null) {
             return new JsonResponse(['error' => 'Cart already exists'], Response::HTTP_BAD_REQUEST);
         }
-        $cart = $this->cartFactory->create($cartDTO, $user);
+        if (($errors = $this->cartService->validateDTO($cartDTO)) !== null) {
+            return $errors;
+        }
 
-        $this->cartRepository->save($cart);
+        $cart = $this->cartFactory->create($cartDTO, $user);
+        $this->entityManager->flush();
 
         return new JsonResponse(['cartId' => $cart->getId()], Response::HTTP_OK);
     }
@@ -65,17 +75,39 @@ class CartController extends AbstractController
     #[Security(name: 'Bearer')]
     public function delete(#[MapEntity(id: 'id')] Cart $cart): JsonResponse
     {
-        if ($cart->getUser() !== $this->getUser()) {
-            if ($this->isGranted('ROLE_ADMIN') === false) {
-                return new JsonResponse(
-                    ['error' => 'You do not have sufficient permissions to perform this action.'],
-                    Response::HTTP_FORBIDDEN,
-                );
-            }
+        $errors = $this->cartService->checkPermission($cart->getUser(), $this->getUser(), $this->isGranted('ROLE_ADMIN'));
+        if ($errors !== null) {
+            return $errors;
         }
         $this->entityManager->remove($cart);
         $this->entityManager->flush();
 
         return new JsonResponse(['message' => 'Successfully deleted'], Response::HTTP_OK);
+    }
+
+    #[Route('/api/carts/{id}', name: 'cart_update', methods: ['PUT'])]
+    #[OA\Put(summary: 'Update cart')]
+    #[IsGranted('IS_AUTHENTICATED_FULLY', message: 'Only auth user can delete cart.')]
+    #[Security(name: 'Bearer')]
+    public function update(
+        #[MapRequestPayload(
+            acceptFormat: 'json',
+            validationFailedStatusCode: Response::HTTP_BAD_REQUEST,
+        )] CartDTO $cartDTO,
+        #[MapEntity(id: 'id')] Cart $cart,
+    ): JsonResponse {
+        if (($errors = $this->cartService->validateDTO($cartDTO)) !== null) {
+            return $errors;
+        }
+        $errors = $this->cartService->checkPermission($cart->getUser(), $this->getUser(), $this->isGranted('ROLE_ADMIN'));
+        if ($errors !== null) {
+            return $errors;
+        }
+
+        $cart->syncWithDTO($cartDTO);
+        $this->entityManager->flush();
+        $cartData = $this->serializer->serialize($cart, 'json', ['groups' => 'cart']);
+
+        return new JsonResponse($cartData, Response::HTTP_OK);
     }
 }
